% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{cloud_storage_authenticate}
\alias{cloud_storage_authenticate}
\title{Authenticate to a Cloud Storage Provider}
\usage{
cloud_storage_authenticate(provider, options)
}
\arguments{
\item{provider}{A character string specifying the cloud provider. Currently
supports "gcs" for Google Cloud Storage or "aws" for Amazon Web Services.}

\item{options}{A named list of options specific to the cloud provider. For
GCS, this should include \code{service_account_key} with the contents of the
authentication JSON file from your Google Project.}
}
\value{
NULL (called for side effects)
}
\description{
This function authenticates to a cloud storage provider using the specified
authentication method. Currently supports Google Cloud Storage (GCS) and
Amazon Web Services (AWS).
}
\examples{
\dontrun{
# Authenticate with Google Cloud Storage
authentication_details <- readLines("path/to/json_file.json")
cloud_storage_authenticate("gcs", list(service_account_key = authentication_details))
}
}
\keyword{storage}
