% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{download_parquet_from_cloud}
\alias{download_parquet_from_cloud}
\title{Download Parquet File from Cloud Storage}
\usage{
download_parquet_from_cloud(prefix, provider, options)
}
\arguments{
\item{prefix}{A character string specifying the file prefix path in cloud storage.}

\item{provider}{A character string specifying the cloud storage provider key.}

\item{options}{A named list of cloud storage provider options.}
}
\value{
A tibble containing the data from the Parquet file.
}
\description{
This function downloads a Parquet file from cloud storage and loads it as a data frame.
}
\examples{
\dontrun{
# Download survey data
data <- download_parquet_from_cloud(
  prefix = "raw-data/survey-data",
  provider = conf$storage$google$key,
  options = conf$storage$google$options
)
}

}
\keyword{storage}
