% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_landings}
\alias{export_landings}
\title{Export Processed Landings Data}
\usage{
export_landings()
}
\value{
None (invisible). Data is exported directly to MongoDB.
}
\description{
Exports validated landings data to a MongoDB collection. This function filters and
transforms preprocessed data, calculates total catch weights, and exports selected
fields to a validated collection.
}
\details{
The function performs the following steps:
\enumerate{
\item Pulls preprocessed data from MongoDB
\item Calculates total catch weights per submission
\item Filters for valid survey activities (survey_activity == 1)
\item Selects relevant fields for export
\item Uploads the filtered data to the validated collection
}
}
\note{
Only submissions with survey_activity == 1 are included in the export.
Catch weights are summed per submission, with NA values removed.
}
\section{Exported Fields}{

The following fields are included in the export:
\itemize{
\item submission_id: Unique identifier for the submission
\item landing_date: Date and time of landing
\item district: Administrative district
\item landing_site: Name of landing site
\item catch_outcome: Outcome of catch
\item lat: Latitude
\item lon: Longitude
\item habitat: Fishing habitat
\item vessel_type: Type of fishing vessel
\item propulsion_gear: Type of propulsion
\item trip_duration: Duration of fishing trip
\item gear: Fishing gear used
\item catch_df: Nested catch data
}
}

\examples{
\dontrun{
export_landings()
}

}
\seealso{
\code{\link{preprocess_landings_adnap}} for data preprocessing
\code{\link{calculate_catch_adnap}} for catch weight calculations
}
\keyword{export}
\keyword{workflow}
