% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_lurio_landings}
\alias{export_lurio_landings}
\title{Export Lurio landings data}
\usage{
export_lurio_landings()
}
\description{
Processes validated landings data and exports multiple summary datasets to MongoDB.
Calculates fishery performance metrics (CPUE, RPUE), aggregates data by site and taxa,
and creates visualization-ready formats for the monitoring portal.
}
\details{
Exports the following collections to MongoDB:
\itemize{
\item monthly_metrics: Time series of catch and revenue metrics by district
\item sites_stats: Landing site statistics and performance indicators
\item taxa_length: Length frequency distributions by species
\item taxa_sites: Top 5 taxa composition per landing site
\item habitat_gears_list: CPUE/RPUE metrics by habitat and gear type
}
}
\examples{
\dontrun{
export_lurio_landings()
}

}
\keyword{export}
\keyword{workflow}
