% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{get_catch_bounds_taxon}
\alias{get_catch_bounds_taxon}
\title{Get Catch Bounds by Gear + Taxon}
\usage{
get_catch_bounds_taxon(data, k)
}
\arguments{
\item{data}{A data frame that includes the columns \code{gear}, \code{catch_taxon},
and \code{catch_kg}.}

\item{k}{A numeric parameter passed to \code{\link[univOutl]{LocScaleB}}, controlling
how "wide" the outlier threshold will be (often \code{k = 3} or \code{k = 2}).}
}
\value{
A data frame with columns \code{gear}, \code{catch_taxon}, \code{upper.up},
and the bounds from \code{LocScaleB}. The \code{lower.low} column is dropped.
}
\description{
Computes upper/lower outlier bounds for catch data (\code{catch_kg}) grouped by
the interaction of \code{gear} and \code{catch_taxon}. This function uses
\code{\link[univOutl]{LocScaleB}} with a (default) log transform to find outlier
thresholds for each \code{gear-taxont} group, then exponentiates the upper bound.
}
\details{
\enumerate{
\item Filters to columns \code{gear}, \code{catch_taxon}, and \code{catch_kg}.
\item Splits the data by the interaction of \code{gear} and \code{catch_taxon}.
\item Runs \code{\link[univOutl]{LocScaleB}} on each subset (with \code{logt = TRUE}),
retrieving the \code{bounds}.
\item Binds the subset results, exponentiates the \code{upper.up} bound, then separates
\code{gear_taxon} back into \code{gear} and \code{catch_taxon}.
}
}
\examples{
\dontrun{
data_bounds <- get_catch_bounds_taxon(data, k = 3)
head(data_bounds)
}

}
\seealso{
\code{\link[univOutl]{LocScaleB}} for outlier detection,
\code{\link{validate_catch_taxa}} for usage of these bounds in a validation step.
}
\keyword{validation}
