% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{ingest_landings}
\alias{ingest_landings}
\title{Download and Process Surveys from Kobotoolbox}
\usage{
ingest_landings(
  url = NULL,
  project_id = NULL,
  username = NULL,
  psswd = NULL,
  encoding = NULL
)
}
\arguments{
\item{url}{The URL of Kobotoolbox (default is NULL, uses value from configuration).}

\item{project_id}{The asset ID of the project to download data from (default is NULL, uses value from configuration).}

\item{username}{Username for Kobotoolbox account (default is NULL, uses value from configuration).}

\item{psswd}{Password for Kobotoolbox account (default is NULL, uses value from configuration).}

\item{encoding}{Encoding to be used for data retrieval (default is NULL, uses "UTF-8").}
}
\value{
No return value. Function downloads data, processes it, and uploads to MongoDB.
}
\description{
This function retrieves survey data from Kobotoolbox for a specific project,
processes it, and uploads the raw data to a MongoDB database. It uses the
\code{get_kobo_data} function, which is a wrapper for \code{kobotools_kpi_data} from
the KoboconnectR package.
}
\details{
The function performs the following steps:
\enumerate{
\item Reads configuration settings.
\item Downloads survey data from Kobotoolbox using \code{get_kobo_data}.
\item Checks for uniqueness of submissions.
\item Converts data to tabular format.
\item Uploads raw data to MongoDB.
}

Note that while parameters are provided for customization, the function
currently uses hardcoded values and configuration settings for some parameters.
}
\examples{
\dontrun{
ingest_surveys(
  url = "eu.kobotoolbox.org",
  project_id = "my_project_id",
  username = "admin",
  psswd = "admin",
  encoding = "UTF-8"
)
}
}
\keyword{ingestion}
\keyword{workflow}
