% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{ingest_wf_surveys}
\alias{ingest_wf_surveys}
\title{Download and Process WF Catch Surveys from Kobotoolbox}
\usage{
ingest_wf_surveys(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{Logging threshold level (default: logger::DEBUG)}
}
\value{
No return value. Function downloads data, processes it, and uploads to Google Cloud Storage.
}
\description{
This function retrieves WF survey data from Kobotoolbox,
processes it, and uploads the raw data as Parquet files to Google Cloud Storage.
This function retrieves WCS survey data  from Kobotoolbox,
processes it, and uploads the raw data as Parquet files to Google Cloud Storage.
}
\details{
The function performs the following steps:
\enumerate{
\item Reads configuration settings.
\item Downloads survey data from Kobotoolbox using \code{get_kobo_data}.
\item Checks for uniqueness of submissions.
\item Converts data to tabular format.
\item Uploads raw data as Parquet files to Google Cloud Storage.
\item Uploads raw data as Parquet files to Google Cloud Storage.
}

This function processes WF surveys.
}
\examples{
\dontrun{
ingest_wf_surveys()
}
}
\keyword{ingestion}
\keyword{workflow}
