% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lw_coeffs}
\alias{lw_coeffs}
\title{Length-weight coefficients}
\format{
A data frame
}
\usage{
data(lw_coeffs)
}
\description{
Length-weight coefficients median for fish groups
}
\keyword{datasets}
