% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peskas.mozambique.data.pipeline-package.R
\docType{package}
\name{peskas.mozambique.data.pipeline-package}
\alias{peskas.mozambique.data.pipeline}
\alias{peskas.mozambique.data.pipeline-package}
\title{peskas.mozambique.data.pipeline: Mozambique Small-Scale Fisheries Data Pipeline}
\description{
This package implements the data and modelling pipelines underpinning the Peskas system in Mozambique.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://worldfishcenter.github.io/peskas.mozambique.data.pipeline/}
}

}
\author{
\strong{Maintainer}: Lorenzo Longobardi \email{l.longobardi@cgiar.org} (\href{https://orcid.org/0000-0003-3126-7341}{ORCID})

Other contributors:
\itemize{
  \item WorldFish [copyright holder]
}

}
\keyword{internal}
