% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{rename_child}
\alias{rename_child}
\title{Rename Child Elements in Nested Kobotoolbox Data}
\usage{
rename_child(x, i, p)
}
\arguments{
\item{x}{The child element to be renamed.}

\item{i}{The index or name of the child element.}

\item{p}{The name of the parent element.}
}
\value{
A renamed list of child elements.
}
\description{
This internal function renames child elements in nested Kobotoolbox data structures.
}
\keyword{internal}
