% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{upload_parquet_to_cloud}
\alias{upload_parquet_to_cloud}
\title{Upload Data as Parquet File to Cloud Storage}
\usage{
upload_parquet_to_cloud(
  data,
  prefix,
  provider,
  options,
  compression = "lz4",
  compression_level = 12
)
}
\arguments{
\item{data}{A data frame containing the data to be uploaded.}

\item{prefix}{A character string specifying the file prefix path.}

\item{provider}{A character string specifying the cloud storage provider key.}

\item{options}{A named list of cloud storage provider options.}

\item{compression}{A character string specifying compression type. Default is "lz4".}

\item{compression_level}{An integer specifying compression level. Default is 12.}
}
\value{
NULL (called for side effects)
}
\description{
This function uploads a data frame as a Parquet file to cloud storage
with versioning support.
}
\examples{
\dontrun{
# Upload survey data
upload_parquet_to_cloud(
  data = survey_data,
  prefix = "raw-data/survey-data",
  provider = conf$storage$google$key,
  options = conf$storage$google$options
)
}

}
\keyword{storage}
