% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_landings}
\alias{validate_landings}
\title{Validate Fisheries Data}
\usage{
validate_landings(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{Logging threshold level (default: logger::DEBUG)}
}
\value{
This function does not return a value. Instead, it processes the data and
uploads the validated results to a MongoDB collection in the pipeline database.
}
\description{
This function imports and validates preprocessed fisheries data from a MongoDB
collection. It conducts a series of validation checks to ensure data integrity,
including checks on \strong{catch weights} at both the taxon level and aggregated
level, and checks on catch \strong{prices}.
}
\details{
The function performs the following main operations:
\enumerate{
\item Pulls preprocessed landings data from the MongoDB collection specified in the config.
\item Summarizes data by \code{submission_id} and \code{catch_taxon}, computing total
\code{catch_kg}, \code{catch_estimate}, and keeping the first value of certain
metadata fields (e.g., \code{landing_date}, \code{landing_site}, \code{gear}, etc.).
\item Filters out records where \code{catch_outcome == 1} but \code{catch_gr == 0}.
\item Validates:
\itemize{
\item \emph{Taxon-level catch} with \code{\link{validate_catch_taxa}}
\item \emph{Total catch} with \code{\link{validate_total_catch}}
\item \emph{Price} with \code{\link{validate_price}}
}
\item Calculates additional price-based checks, such as \emph{price per kg}.
\item Merges all validation flags into a single dataset and constructs a combined
alert column.
\item Finally, it filters out any records with alerts, leaving only valid data,
and uploads the filtered data back to another MongoDB collection (the
"validated" collection).
}
}
\note{
This function requires a configuration file to be present and readable by
the \code{read_config} function, which should provide MongoDB connection details
and parameters for validation.
}
\examples{
\dontrun{
validate_landings()
}

}
\keyword{validation}
\keyword{workflow}
