% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peskas.mozambique.data.pipeline-package.R
\docType{package}
\name{peskas.mozambique.data.pipeline-package}
\alias{peskas.mozambique.data.pipeline}
\alias{peskas.mozambique.data.pipeline-package}
\title{peskas.mozambique.data.pipeline: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://worldfishcenter.github.io/peskas.mozambique.data.pipeline/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
