% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-surveys.R
\name{process_version_data}
\alias{process_version_data}
\title{Process Version Data Helper Function}
\usage{
process_version_data(catch_info = NULL, asfis = NULL)
}
\arguments{
\item{catch_info}{Processed catch information}

\item{asfis}{ASFIS species data}
}
\value{
Combined and processed survey data
}
\description{
Internal helper function to process catch and general info for a specific survey version
}
\keyword{internal}
